Start refactoring
I still need to find out how to communicate between the
display element and the filterer.
